+++ parser/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
+## Unreleased
+
+### Changed
+
+- Support events with parameters (e.g `{ trigger event(param1, param2, "text", ...) }`)
+
 ## 2.4.0 (2024-10-07)
 
 ### Added